second commit

+++ src/main/java/com/jpa/test/dao/OrderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jpa.test.entites.Orders;
 
-public interface OrderRepository extends JpaRepository<Orders,Integer>{
+public interface OrderRepository extends JpaRepository<Orders,Short>{
 
 }
