+++ src/main/java/com/jpa/test/dao/ProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.jpa.test.entites.Product;
 
 
-public interface ProductRepository extends JpaRepository<Product,Integer>{
+public interface ProductRepository extends JpaRepository<Product,Short>{
 
 }
