+++ src/main/java/com/jpa/test/dao/ShippingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jpa.test.entites.Shipping;
 
-public interface ShippingRepository extends JpaRepository<Shipping,Integer> {
+public interface ShippingRepository extends JpaRepository<Shipping,Short> {
 
 }
